refactor(interp): unify jit plan compilation
