+++ interp/jit_arm64.go
@@ -81,35 +81,50 @@ func newCompiler() (*compiler, error) {
 	}, nil
 }
 
-// lower lowers a recorded entry trace tree into one framed native callable.
-// The hot path follows the recorded opcodes linearly with scalars unboxed in
-// registers; recorded branch directions, observed call targets, and narrow
-// value guards exit through the journal and resume through threaded dispatch.
-// Returns false whenever the trace touches an opcode, kind, or shape the
-// typed lowerer does not support; the caller leaves threaded dispatch installed.
-func (l arm64Lowerer) lower(ctx *lowering) bool {
-	l.enter(ctx)
-	// A loop root re-enters the live frame at the header: bind the back-edge
-	// target after the prologue so each iteration reloads its loop-carried
-	// locals from the VM stack (flushed at the back-edge) without a memory-free
-	// register fixpoint. An entry trace binds it too, so a self tail-call that
-	// loops back to the function entry has a branch target (see tailLoop). The
-	// label is harmless when no back-edge references it.
-	ctx.back = ctx.assembler.Label()
-	ctx.assembler.Bind(ctx.back)
-	if !l.walk(ctx, ctx.tree.root.ops, nil, ctx.tree.root.kind) {
+// lower emits one backend-neutral plan into a framed native callable.
+func (l arm64Lowerer) lower(ctx *lowering, plan plan) bool {
+	if len(plan.blocks) == 0 {
 		return false
 	}
-	for n := 0; n < len(ctx.pending); n++ {
-		sort.SliceStable(ctx.pending[n:], func(i, j int) bool {
-			return ctx.pending[n+i].hits > ctx.pending[n+j].hits
+	l.enter(ctx)
+	ctx.root = plan.blocks[0]
+	ctx.kind = plan.entry.kind
+
+	if ctx.root.canonical {
+		labels := make([]asm.Label, len(plan.blocks))
+		index := make(map[int]int, len(plan.blocks))
+		for idx, block := range plan.blocks {
+			labels[idx] = ctx.assembler.Label()
+			index[block.offset] = idx
+		}
+		for idx, block := range plan.blocks {
+			ctx.assembler.Bind(labels[idx])
+			if !l.emitBlock(ctx, block, nil, labels, index) {
+				return false
+			}
+		}
+		l.materializeExits(ctx)
+		return true
+	}
+
+	for _, block := range plan.blocks[1:] {
+		ctx.branches[branch{fn: block.address, ip: block.offset}] = block
+	}
+	ctx.back = ctx.assembler.Label()
+	ctx.assembler.Bind(ctx.back)
+	if !l.emitBlock(ctx, ctx.root, nil, nil, nil) {
+		return false
+	}
+	for n := 0; n < len(ctx.work); n++ {
+		sort.SliceStable(ctx.work[n:], func(i, j int) bool {
+			return ctx.work[n+i].hits > ctx.work[n+j].hits
 		})
-		p := ctx.pending[n]
-		ctx.values = p.values
-		ctx.frames = p.frames
-		ctx.assembler.Bind(p.label)
+		work := ctx.work[n]
+		ctx.values = work.values
+		ctx.frames = work.frames
+		ctx.assembler.Bind(work.label)
 		l.reload(ctx)
-		if !l.walk(ctx, p.ops, p.tail, p.kind) {
+		if !l.emitBlock(ctx, work.block, work.tail, nil, nil) {
 			return false
 		}
 	}
@@ -152,7 +167,7 @@ func (l arm64Lowerer) materializeExits(ctx *lowering) {
 // partial trace exits at its cut so the next segment can be learned as a hot
 // continuation. A loop trace additionally closes one back-edge: a BR or BR_IF
 // whose target is the loop header becomes the native back-edge (see backedge).
-func (l arm64Lowerer) walk(ctx *lowering, ops, tail []step, kind outcome) bool {
+func (l arm64Lowerer) walk(ctx *lowering, ops, tail []step, kind terminatorKind) bool {
 	for idx := 0; idx < len(ops); idx++ {
 		op := ops[idx]
 		f := ctx.frame()
@@ -170,7 +185,7 @@ func (l arm64Lowerer) walk(ctx *lowering, ops, tail []step, kind outcome) bool {
 			ctx.assembler.Emit(arm64.BLabel(label))
 			return true
 		}
-		if ctx.loop && (op.op == instr.BR || op.op == instr.BR_IF) && op.target == ctx.tree.root.anchor.ip {
+		if ctx.kind == entryLoop && (op.op == instr.BR || op.op == instr.BR_IF) && op.target == ctx.root.offset {
 			if !l.backedge(ctx, op) {
 				return false
 			}
@@ -260,7 +275,7 @@ func (l arm64Lowerer) walk(ctx *lowering, ops, tail []step, kind outcome) bool {
 			instr.F64_LE,
 			instr.F64_GE,
 			instr.ARRAY_GET:
-			handled, emitted := l.emitStep(ctx, op, false)
+			handled, emitted := l.emitStep(ctx, op)
 			ok = handled && emitted
 		case instr.UNREACHABLE:
 			ok = l.unreachable(ctx, op)
@@ -536,9 +551,9 @@ func (l arm64Lowerer) walk(ctx *lowering, ops, tail []step, kind outcome) bool {
 		if !ok {
 			label = ctx.assembler.Label()
 			ctx.tails[&tail[0]] = label
-			p := pending{label: label, ops: tail}
+			p := work{label: label, block: block{steps: tail}}
 			p.values, p.frames = ctx.snapshot()
-			ctx.pending = append(ctx.pending, p)
+			ctx.work = append(ctx.work, p)
 		}
 		ctx.assembler.Emit(arm64.BLabel(label))
 		return true
@@ -551,7 +566,7 @@ func (l arm64Lowerer) walk(ctx *lowering, ops, tail []step, kind outcome) bool {
 	// end) or returned (e.g. a nested inlined callee's RETURN popped back to
 	// the caller, handled by the RETURN case's stitch()+break above, leaving
 	// the caller's remaining ops as the fragment's natural end). A tail
-	// fragment's zero-value kind (see the pending literal above, which
+	// fragment's zero-value kind (see the work item above, which
 	// doesn't set kind) is likewise not aborted and may fall through.
 	//
 	// Only aborted must never fall through: it recorded a prefix and hit an
@@ -562,10 +577,12 @@ func (l arm64Lowerer) walk(ctx *lowering, ops, tail []step, kind outcome) bool {
 	// fragment's own kind was the bug this replaced: a side-exit
 	// continuation that itself aborted could be lowered as if it had
 	// completed normally, because the entry root's kind is completed.
-	if kind != aborted {
+	switch kind {
+	case terminateFallthrough, terminateReturn, terminateComplete:
 		return l.complete(ctx)
+	default:
+		return false
 	}
-	return false
 }
 
 // fuse lowers short sequences whose intermediate ref ownership can be
@@ -623,7 +640,7 @@ func (l arm64Lowerer) consume(ctx *lowering, ops, tail []step, idx int) (int, bo
 	consumed := 2
 	if consumer.op == instr.REF_IS_NULL && idx+2 < len(ops) {
 		branch := ops[idx+2]
-		if l.adjacent(consumer, branch) && branch.op == instr.BR_IF && (!ctx.loop || branch.target != ctx.tree.root.anchor.ip) {
+		if l.adjacent(consumer, branch) && branch.op == instr.BR_IF && (ctx.kind != entryLoop || branch.target != ctx.root.offset) {
 			consumed = 3
 		}
 	}
@@ -1742,7 +1759,7 @@ func (l arm64Lowerer) sign64(ctx *lowering, v asm.VReg) asm.VReg {
 // so it falls back instead of ending at the callee RETURN.
 //
 // ops is only the fragment being walked; tail is that fragment's own
-// continuation — the caller path a pending leg still owes. When the current
+// continuation — the caller path a deferred block still owes. When the current
 // frame's return is not represented inside ops (every remaining step sits at the
 // branch's depth or deeper), the resume point lives in tail, so a branch nested
 // inside a learned leg must inherit tail rather than truncate at the leg's end.
@@ -1877,23 +1894,22 @@ func (l arm64Lowerer) branchClean(ctx *lowering, ip int, tail []step) (asm.Label
 
 // continuation materializes the current symbolic state and returns a native
 // label for a learned branch target, or false when the target is not eligible
-// (no learned trace, a marked stack, or too many pending continuations).
+// (no learned trace, a marked stack, or too many deferred continuations).
 // Flushed callers write VM stack homes before jumping to the reloading branch
 // body; clean callers skip the no-op flush after proving those homes are
-// already current. Caller-tailed continuations compile as distinct pending
+// already current. Caller-tailed continuations compile as distinct deferred
 // blocks because the same branch target may need different caller remainders.
 func (l arm64Lowerer) continuation(ctx *lowering, ip int, tail []step, flush bool) (asm.Label, bool) {
 	target := branch{fn: ctx.frame().addr, ip: ip}
-	leg := ctx.branches[target]
+	leg, ok := ctx.branches[target]
 	// Loop roots require the loop lowering mode: it turns their recorded
 	// back-edge into a real native back-edge with a safepoint budget. Inlining
 	// one as an ordinary continuation would instead treat the recorded path as
 	// top-level completion and return after a single iteration. Exit to the
 	// threaded loop header so its independently compiled loop entry runs.
-	// An aborted leg recorded only a partial, unsupported fragment (already
-	// excluded from ctx.branches by tree.branchIPs); the check is repeated
-	// here as a direct safeguard against inlining it as a continuation.
-	if leg.trace == nil || leg.trace.kind == loop || leg.trace.kind == aborted || l.marked(ctx) || ctx.pendingBranches >= pendingLimit {
+	// The planner excludes loop roots and aborted fragments, so this path only
+	// schedules complete, returned, or resumable continuation blocks.
+	if !ok || l.marked(ctx) || ctx.scheduled >= continuationLimit {
 		return 0, false
 	}
 	if flush && !l.flush(ctx, false) {
@@ -1908,10 +1924,10 @@ func (l arm64Lowerer) continuation(ctx *lowering, ip int, tail []step, flush boo
 	if len(tail) == 0 {
 		ctx.queued[target] = label
 	}
-	p := pending{label: label, ops: leg.trace.ops, tail: tail, kind: leg.trace.kind, hits: leg.hits}
+	p := work{label: label, block: leg, tail: tail, hits: leg.hits}
 	p.values, p.frames = ctx.snapshot()
-	ctx.pending = append(ctx.pending, p)
-	ctx.pendingBranches++
+	ctx.work = append(ctx.work, p)
+	ctx.scheduled++
 	return label, true
 }
 
@@ -2387,7 +2403,7 @@ func (l arm64Lowerer) sideExit(ctx *lowering, pre []value, resume int) (asm.Labe
 // carried locals are committed to the VM stack before any split so both the
 // looping and the exit path read them back consistently.
 func (l arm64Lowerer) backedge(ctx *lowering, op step) bool {
-	header := ctx.tree.root.anchor.ip
+	header := ctx.root.offset
 	a := ctx.assembler
 	if op.op == instr.BR_IF {
 		if ctx.count() < 1 || !l.kinds(ctx, types.KindI32, 1) {
@@ -3585,7 +3601,7 @@ func (l arm64Lowerer) args(ctx *lowering, target *types.Function, params int) bo
 	return true
 }
 
-// marked reports whether a raw ref marker blocks pending continuation reload.
+// marked reports whether a raw ref marker blocks deferred continuation reload.
 func (l arm64Lowerer) marked(ctx *lowering) bool {
 	for _, v := range ctx.values {
 		if v.kind == types.KindRef && v.raw {
@@ -3638,4 +3654,589 @@ func (l arm64Lowerer) zero32(ctx *lowering, v asm.VReg) asm.VReg {
 
 func (arm64Lowerer) narrow32(v asm.VReg) asm.VReg {
 	return asm.NewVReg(v.ID(), v.Type(), asm.Width32)
+}
+
+func (l arm64Lowerer) emitStep(ctx *lowering, op step) (bool, bool) {
+	ok := false
+	switch op.op {
+	case instr.I32_CONST:
+		ok = l.i32Const(ctx, op)
+	case instr.I64_CONST:
+		ok = l.i64Const(ctx, op)
+	case instr.F32_CONST:
+		ok = l.f32Const(ctx, op)
+	case instr.F64_CONST:
+		ok = l.f64Const(ctx, op)
+	case instr.CONST_GET:
+		if op.known {
+			ok = l.constGetKnown(ctx, op)
+		} else {
+			ok = l.constGet(ctx, op)
+		}
+	case instr.LOCAL_GET:
+		ok = l.localGet(ctx, op)
+	case instr.LOCAL_SET:
+		ok = l.localSet(ctx, op, true)
+	case instr.LOCAL_TEE:
+		ok = l.localSet(ctx, op, false)
+	case instr.GLOBAL_GET:
+		ok = l.globalGet(ctx, op)
+	case instr.GLOBAL_SET:
+		ok = l.globalSet(ctx, op, true)
+	case instr.GLOBAL_TEE:
+		ok = l.globalSet(ctx, op, false)
+	case instr.DROP:
+		ok = l.drop(ctx, op)
+	case instr.DUP:
+		ok = l.dup(ctx)
+	case instr.SWAP:
+		ok = l.swap(ctx)
+	case instr.SELECT:
+		ok = l.selectOp(ctx)
+	case instr.I32_ADD:
+		ok = l.i32Binary(ctx, arm64.ADD)
+	case instr.I32_SUB:
+		ok = l.i32Binary(ctx, arm64.SUB)
+	case instr.I32_MUL:
+		ok = l.i32Binary(ctx, arm64.MUL)
+	case instr.I32_AND:
+		ok = l.i32Bitwise(ctx, arm64.AND)
+	case instr.I32_OR:
+		ok = l.i32Bitwise(ctx, arm64.ORR)
+	case instr.I32_XOR:
+		ok = l.i32Bitwise(ctx, arm64.EOR)
+	case instr.I32_EQZ:
+		ok = l.i32Eqz(ctx)
+	case instr.I32_EQ:
+		ok = l.i32Cmp(ctx, arm64.CondEQ)
+	case instr.I32_NE:
+		ok = l.i32Cmp(ctx, arm64.CondNE)
+	case instr.I32_LT_S:
+		ok = l.i32Cmp(ctx, arm64.CondLT)
+	case instr.I32_LE_S:
+		ok = l.i32Cmp(ctx, arm64.CondLE)
+	case instr.I32_GT_S:
+		ok = l.i32Cmp(ctx, arm64.CondGT)
+	case instr.I32_GE_S:
+		ok = l.i32Cmp(ctx, arm64.CondGE)
+	case instr.I32_LT_U:
+		ok = l.i32Cmp(ctx, arm64.CondCC)
+	case instr.I32_LE_U:
+		ok = l.i32Cmp(ctx, arm64.CondLS)
+	case instr.I32_GT_U:
+		ok = l.i32Cmp(ctx, arm64.CondHI)
+	case instr.I32_GE_U:
+		ok = l.i32Cmp(ctx, arm64.CondCS)
+	case instr.I64_ADD:
+		ok = l.i64Binary(ctx, op, arm64.ADD, true)
+	case instr.I64_SUB:
+		ok = l.i64Binary(ctx, op, arm64.SUB, true)
+	case instr.I64_MUL:
+		ok = l.i64Binary(ctx, op, arm64.MUL, true)
+	case instr.I64_AND:
+		ok = l.i64Binary(ctx, op, arm64.AND, false)
+	case instr.I64_OR:
+		ok = l.i64Binary(ctx, op, arm64.ORR, false)
+	case instr.I64_XOR:
+		ok = l.i64Binary(ctx, op, arm64.EOR, false)
+	case instr.I64_EQZ:
+		ok = l.i64Eqz(ctx)
+	case instr.I64_EQ:
+		ok = l.i64Cmp(ctx, arm64.CondEQ)
+	case instr.I64_NE:
+		ok = l.i64Cmp(ctx, arm64.CondNE)
+	case instr.I64_LT_S:
+		ok = l.i64Cmp(ctx, arm64.CondLT)
+	case instr.I64_LE_S:
+		ok = l.i64Cmp(ctx, arm64.CondLE)
+	case instr.I64_GT_S:
+		ok = l.i64Cmp(ctx, arm64.CondGT)
+	case instr.I64_GE_S:
+		ok = l.i64Cmp(ctx, arm64.CondGE)
+	case instr.I64_LT_U:
+		ok = l.i64Cmp(ctx, arm64.CondCC)
+	case instr.I64_LE_U:
+		ok = l.i64Cmp(ctx, arm64.CondLS)
+	case instr.I64_GT_U:
+		ok = l.i64Cmp(ctx, arm64.CondHI)
+	case instr.I64_GE_U:
+		ok = l.i64Cmp(ctx, arm64.CondCS)
+	case instr.F32_ADD:
+		ok = l.f32Binary(ctx, arm64.FADD)
+	case instr.F32_SUB:
+		ok = l.f32Binary(ctx, arm64.FSUB)
+	case instr.F32_MUL:
+		ok = l.f32Binary(ctx, arm64.FMUL)
+	case instr.F32_DIV:
+		ok = l.f32Binary(ctx, arm64.FDIV)
+	case instr.F32_EQ:
+		ok = l.f32Cmp(ctx, arm64.CondEQ)
+	case instr.F32_NE:
+		ok = l.f32Cmp(ctx, arm64.CondNE)
+	case instr.F32_LT:
+		ok = l.f32Cmp(ctx, arm64.CondMI)
+	case instr.F32_GT:
+		ok = l.f32Cmp(ctx, arm64.CondGT)
+	case instr.F32_LE:
+		ok = l.f32Cmp(ctx, arm64.CondLS)
+	case instr.F32_GE:
+		ok = l.f32Cmp(ctx, arm64.CondGE)
+	case instr.F64_ADD:
+		ok = l.f64Binary(ctx, arm64.FADD)
+	case instr.F64_SUB:
+		ok = l.f64Binary(ctx, arm64.FSUB)
+	case instr.F64_MUL:
+		ok = l.f64Binary(ctx, arm64.FMUL)
+	case instr.F64_DIV:
+		ok = l.f64Binary(ctx, arm64.FDIV)
+	case instr.F64_EQ:
+		ok = l.f64Cmp(ctx, arm64.CondEQ)
+	case instr.F64_NE:
+		ok = l.f64Cmp(ctx, arm64.CondNE)
+	case instr.F64_LT:
+		ok = l.f64Cmp(ctx, arm64.CondMI)
+	case instr.F64_GT:
+		ok = l.f64Cmp(ctx, arm64.CondGT)
+	case instr.F64_LE:
+		ok = l.f64Cmp(ctx, arm64.CondLS)
+	case instr.F64_GE:
+		ok = l.f64Cmp(ctx, arm64.CondGE)
+	case instr.ARRAY_GET:
+		if ctx.count() >= 2 && ctx.values[len(ctx.values)-2].raw && ctx.values[len(ctx.values)-2].ref > 0 {
+			ok = l.arrayGetKnown(ctx, op)
+		} else {
+			ok = l.arrayGet(ctx, op)
+		}
+	default:
+		return false, false
+	}
+	return true, ok
+}
+
+func (l arm64Lowerer) constGetKnown(ctx *lowering, op step) bool {
+	idx := int(instr.Instruction(ctx.frame().code[op.ip:]).Operand(0))
+	if idx >= len(ctx.constants) {
+		return false
+	}
+	boxed := ctx.constants[idx]
+	if boxed.Kind() != types.KindRef {
+		return l.constGet(ctx, op)
+	}
+	ref := boxed.Ref()
+	if ref <= 0 || ref >= len(ctx.heap) {
+		return false
+	}
+	switch ctx.heap[ref].(type) {
+	case types.TypedArray[bool], types.TypedArray[int8], types.TypedArray[int32],
+		types.TypedArray[float32], types.TypedArray[float64]:
+		ctx.push(value{kind: types.KindRef, raw: true, ref: ref})
+		return true
+	default:
+		return l.constGet(ctx, op)
+	}
+}
+
+func (l arm64Lowerer) arrayGetKnown(ctx *lowering, op step) bool {
+	if ctx.count() < 2 || ctx.values[len(ctx.values)-1].kind != types.KindI32 {
+		return false
+	}
+	marker := ctx.values[len(ctx.values)-2]
+	constant := marker.ref
+	if !marker.raw || constant <= 0 || constant >= len(ctx.heap) {
+		return false
+	}
+
+	var kind types.Kind
+	var want uintptr
+	var scale uint8
+	switch value := ctx.heap[constant].(type) {
+	case types.TypedArray[bool]:
+		kind, want = types.KindI1, itab(value)
+	case types.TypedArray[int8]:
+		kind, want = types.KindI8, itab(value)
+	case types.TypedArray[int32]:
+		kind, want, scale = types.KindI32, itab(value), 2
+	case types.TypedArray[float32]:
+		kind, want, scale = types.KindF32, itab(value), 2
+	case types.TypedArray[float64]:
+		kind, want, scale = types.KindF64, itab(value), 3
+	default:
+		return false
+	}
+
+	pre := append([]value(nil), ctx.values...)
+	boxed := ctx.assembler.Reg(asm.RegTypeInt, asm.Width64)
+	ctx.assembler.Emit(arm64.LDI(boxed, uint64(types.BoxRef(constant)))...)
+	ctx.values[len(ctx.values)-2] = value{reg: boxed, kind: types.KindRef, raw: false, ref: constant}
+	if !l.flush(ctx, false) {
+		return false
+	}
+	clear(ctx.frame().loaded)
+	clear(ctx.frame().dirty)
+	fail := ctx.queueExit(nil, op.ip, constant)
+
+	a := ctx.assembler
+	heap := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.LDR(heap, ctx.pin(scratchCtrl), int16(journalHeap*8)))
+	off := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.LDI(off, uint64(constant))...)
+	a.Emit(arm64.LSLI(off, off, 4))
+	cell := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.ADD(cell, heap, off))
+	actual := a.Reg(asm.RegTypeInt, asm.Width64)
+	data := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.LDR(actual, cell, 0), arm64.LDR(data, cell, 8))
+	l.guardItab(ctx, actual, want, fail)
+
+	idx := l.sign32(ctx, ctx.values[len(ctx.values)-1].reg)
+	dataPtr, n := l.sliceHeader(ctx, data, 0)
+	l.guardIndex(ctx, idx, n, fail)
+	result := a.Reg(asm.RegTypeInt, asm.Width64)
+	switch kind {
+	case types.KindI1:
+		elemAddr := a.Reg(asm.RegTypeInt, asm.Width64)
+		a.Emit(arm64.ADD(elemAddr, dataPtr, idx))
+		a.Emit(arm64.LDRB(result, elemAddr, 0))
+	case types.KindI8:
+		elemAddr := a.Reg(asm.RegTypeInt, asm.Width64)
+		elem := a.Reg(asm.RegTypeInt, asm.Width64)
+		a.Emit(arm64.ADD(elemAddr, dataPtr, idx))
+		a.Emit(arm64.LDRB(elem, elemAddr, 0))
+		a.Emit(arm64.SXTB(result, elem))
+	case types.KindI32, types.KindF32:
+		elemOff := a.Reg(asm.RegTypeInt, asm.Width64)
+		elemAddr := a.Reg(asm.RegTypeInt, asm.Width64)
+		a.Emit(arm64.LSLI(elemOff, idx, scale))
+		a.Emit(arm64.ADD(elemAddr, dataPtr, elemOff))
+		a.Emit(arm64.LDRSW(result, elemAddr, 0))
+	case types.KindF64:
+		elemOff := a.Reg(asm.RegTypeInt, asm.Width64)
+		elemAddr := a.Reg(asm.RegTypeInt, asm.Width64)
+		a.Emit(arm64.LSLI(elemOff, idx, scale))
+		a.Emit(arm64.ADD(elemAddr, dataPtr, elemOff))
+		a.Emit(arm64.LDR(result, elemAddr, 0))
+	}
+	ctx.values = append(pre[:len(pre)-2:len(pre)-2], value{reg: result, kind: kind, raw: true})
+	return true
+}
+
+const branchTableLimit = 32
+
+func (l arm64Lowerer) enterBlock(ctx *lowering, state state) {
+	ctx.values = ctx.values[:0]
+	for _, slot := range state.slots {
+		ctx.values = append(ctx.values, value{kind: slot.kind, ref: slot.ref, raw: slot.refKnown})
+	}
+	frame := ctx.frame()
+	clear(frame.loaded)
+	clear(frame.dirty)
+	l.reload(ctx)
+}
+
+func (l arm64Lowerer) emitBlock(ctx *lowering, block block, tail []step, labels []asm.Label, index map[int]int) bool {
+	if !block.canonical {
+		return l.walk(ctx, block.steps, tail, block.term.kind)
+	}
+	l.enterBlock(ctx, block.state)
+	for idx := 0; idx < len(block.steps); idx++ {
+		op := block.steps[idx]
+		if op.op == instr.CONST_GET && idx+1 < len(block.steps) && block.steps[idx+1].op == instr.CALL {
+			inst := instr.Instruction(ctx.frame().code[op.ip:])
+			callee, ok := l.directTarget(ctx, inst)
+			if !ok {
+				return l.exit(ctx, op.ip)
+			}
+			call := block.steps[idx+1]
+			call.callee = callee
+			if !l.directCall(ctx, call) {
+				return false
+			}
+			idx++
+			continue
+		}
+		handled, ok := l.emitStep(ctx, op)
+		if !handled {
+			return l.exit(ctx, op.ip)
+		}
+		if !ok {
+			return false
+		}
+	}
+
+	switch block.term.kind {
+	case terminateBranch:
+		if len(block.term.targets) != 1 || !l.flush(ctx, false) {
+			return false
+		}
+		return l.emitEdge(ctx, block.term.targets[0], block.offset, labels, index)
+	case terminateBranchIf:
+		if len(block.term.targets) != 2 {
+			return false
+		}
+		return l.emitBranchIf(ctx, step{op: instr.BR_IF, fn: ctx.addr, ip: block.term.ip}, block.offset, block.term.targets[1], labels, index)
+	case terminateBranchTable:
+		return l.emitBranchTable(ctx, step{op: instr.BR_TABLE, fn: ctx.addr, ip: block.term.ip}, block.offset, labels, index)
+	case terminateReturn:
+		return l.ret(ctx)
+	case terminateComplete:
+		return l.complete(ctx)
+	case terminateFallback:
+		return l.exit(ctx, block.term.ip)
+	default:
+		return false
+	}
+}
+
+func (l arm64Lowerer) directTarget(ctx *lowering, inst instr.Instruction) (int, bool) {
+	idx := int(inst.Operand(0))
+	if idx >= len(ctx.constants) || ctx.constants[idx].Kind() != types.KindRef {
+		return 0, false
+	}
+	ref := ctx.constants[idx].Ref()
+	if ref <= 0 || ref >= len(ctx.heap) {
+		return 0, false
+	}
+	fn, ok := ctx.heap[ref].(*types.Function)
+	if !ok || fn.Typ == nil || len(fn.Captures) != 0 {
+		return 0, false
+	}
+	ctx.push(value{kind: types.KindRef, raw: true, fn: ref, ref: ref})
+	return ref, true
+}
+
+func (l arm64Lowerer) directCall(ctx *lowering, op step) bool {
+	target := ctx.funcs[op.callee]
+	if op.callee == ctx.addr || target == nil || target.Typ == nil || ctx.count() < 1 {
+		return false
+	}
+	params := len(target.Typ.Params)
+	if ctx.count() < params+1 {
+		return false
+	}
+	for _, typ := range target.Typ.Returns {
+		switch typ.Kind() {
+		case types.KindI1, types.KindI8, types.KindI32, types.KindI64, types.KindF32, types.KindF64:
+		default:
+			return false
+		}
+	}
+
+	a := ctx.assembler
+	vCtrl := ctx.pin(scratchCtrl)
+	natives := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.LDR(natives, vCtrl, int16(journalNatives*8)))
+	callee := a.Reg(asm.RegTypeInt, asm.Width64)
+	if op.callee > 4095 {
+		return false
+	}
+	a.Emit(arm64.LDR(callee, natives, int16(op.callee*8)))
+	ready := a.Label()
+	a.Emit(arm64.CBNZLabel(callee, ready))
+	if !l.exit(ctx, op.ip) {
+		return false
+	}
+	a.Bind(ready)
+
+	marker := ctx.pop()
+	if marker.fn != op.callee || ctx.count() < params || !l.args(ctx, target, params) {
+		return false
+	}
+	if !l.flush(ctx, false) {
+		return false
+	}
+
+	active := ctx.pinTo(arm64.X15)
+	limit := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.LDR(limit, vCtrl, int16(journalCap*8)))
+	a.Emit(arm64.CMP(active, limit))
+	hasFrame := a.Label()
+	a.Emit(arm64.BCondLabel(arm64.OpBCC, hasFrame))
+	l.overflow(ctx, op)
+	a.Bind(hasFrame)
+	a.Emit(arm64.ADDI(active, active, 1))
+	a.Emit(arm64.STR(active, vCtrl, int16(journalActive*8)))
+
+	vBP := ctx.pin(scratchBP)
+	nextSP := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.ADDI(nextSP, vBP, uint16(ctx.sp())))
+	oldBP := ctx.scratch[scratchBP]
+	oldSP := ctx.scratch[scratchSP]
+	a.Emit(
+		arm64.SUBI(arm64.SP, arm64.SP, 32),
+		arm64.STP(oldBP, oldSP, arm64.SP, 0),
+		arm64.STR(arm64.LR, arm64.SP, 16),
+	)
+	calleeBP := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.SUBI(calleeBP, nextSP, uint16(params)))
+	a.Emit(arm64.MOV(ctx.pinTo(oldBP), calleeBP))
+
+	localKinds := target.LocalKinds()
+	if len(localKinds) > params {
+		stack := ctx.pin(scratchStack)
+		base := l.base(ctx, stack)
+		for idx := params; idx < len(localKinds); idx++ {
+			zero, ok := zeroValue(localKinds[idx])
+			if !ok {
+				return false
+			}
+			reg := a.Reg(asm.RegTypeInt, asm.Width64)
+			a.Emit(arm64.LDI(reg, uint64(zero))...)
+			a.Emit(arm64.STR(reg, base, int16((ctx.sp()-params+idx)*8)))
+		}
+	}
+	calleeSP := calleeBP
+	if len(localKinds) > 0 {
+		calleeSP = a.Reg(asm.RegTypeInt, asm.Width64)
+		a.Emit(arm64.ADDI(calleeSP, calleeBP, uint16(len(localKinds))))
+	}
+	a.Emit(arm64.MOV(ctx.pinTo(oldSP), calleeSP))
+	a.Emit(arm64.MOV(arm64.X0, vCtrl))
+	a.Emit(arm64.BLR(callee))
+
+	vCtrl = ctx.pin(scratchCtrl)
+	trap := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.LDR(trap, vCtrl, int16(journalTrap*8)))
+	normal := a.Label()
+	a.Emit(arm64.CBZLabel(trap, normal), arm64.LDR(oldBP, arm64.SP, 0))
+	l.unwind(ctx, vCtrl, op.ip+1)
+	a.Emit(
+		arm64.LDR(arm64.LR, arm64.SP, 16),
+		arm64.ADDI(arm64.SP, arm64.SP, 32),
+		arm64.RET(),
+	)
+	a.Bind(normal)
+
+	active = ctx.pinTo(arm64.X15)
+	a.Emit(arm64.SUBI(active, active, 1))
+	a.Emit(arm64.STR(active, vCtrl, int16(journalActive*8)))
+	a.Emit(
+		arm64.LDP(oldBP, oldSP, arm64.SP, 0),
+		arm64.LDR(arm64.LR, arm64.SP, 16),
+		arm64.ADDI(arm64.SP, arm64.SP, 32),
+	)
+
+	rets := target.Typ.Returns
+	regs := make([]asm.VReg, len(rets))
+	for idx := range rets {
+		if idx >= len(arm64.IntRets) {
+			return false
+		}
+		regs[idx] = ctx.pinTo(arm64.IntRets[idx])
+	}
+	ctx.values = ctx.values[:len(ctx.values)-params]
+	for fi := range ctx.frames {
+		clear(ctx.frames[fi].loaded)
+		clear(ctx.frames[fi].dirty)
+	}
+	l.reload(ctx)
+	for idx, typ := range rets {
+		ctx.push(value{reg: regs[idx], kind: typ.Kind(), raw: true})
+	}
+	return true
+}
+
+func zeroValue(kind types.Kind) (types.Boxed, bool) {
+	switch kind {
+	case types.KindI1:
+		return types.BoxI1(false), true
+	case types.KindI8:
+		return types.BoxI8(0), true
+	case types.KindI32:
+		return types.BoxI32(0), true
+	case types.KindI64:
+		return types.BoxI64(0), true
+	case types.KindF32:
+		return types.BoxF32(0), true
+	case types.KindF64:
+		return types.BoxF64(0), true
+	case types.KindRef:
+		return types.BoxedNull, true
+	default:
+		return 0, false
+	}
+}
+
+func (l arm64Lowerer) emitBranchIf(ctx *lowering, op step, from, nextIP int, labels []asm.Label, index map[int]int) bool {
+	if ctx.count() < 1 || !l.kinds(ctx, types.KindI32, 1) {
+		return false
+	}
+	cond := ctx.pop()
+	if !l.flush(ctx, false) {
+		return false
+	}
+	targets := instr.Targets(ctx.frame().code, op.ip)
+	if len(targets) != 1 {
+		return false
+	}
+	taken := ctx.assembler.Label()
+	ctx.assembler.Emit(arm64.CBNZLabel(l.narrow32(cond.reg), taken))
+	if !l.emitEdge(ctx, nextIP, from, labels, index) {
+		return false
+	}
+	ctx.assembler.Bind(taken)
+	return l.emitEdge(ctx, targets[0], from, labels, index)
+}
+
+func (l arm64Lowerer) emitBranchTable(ctx *lowering, op step, from int, labels []asm.Label, index map[int]int) bool {
+	if ctx.count() < 1 || !l.kinds(ctx, types.KindI32, 1) {
+		return false
+	}
+	code := ctx.frame().code
+	count := int(code[op.ip+1])
+	if count > branchTableLimit {
+		return false
+	}
+	targets := instr.Targets(code, op.ip)
+	if len(targets) != count+1 {
+		return false
+	}
+	cond := ctx.pop()
+	cond32 := ctx.assembler.Reg(asm.RegTypeInt, asm.Width64)
+	ctx.assembler.Emit(arm64.ANDI(cond32, cond.reg, maskI32))
+	if !l.flush(ctx, false) {
+		return false
+	}
+	paths := make([]asm.Label, len(targets))
+	for idx := range paths {
+		paths[idx] = ctx.assembler.Label()
+	}
+	for idx := 0; idx < count; idx++ {
+		ctx.assembler.Emit(arm64.CMPI(cond32, uint16(idx)))
+		ctx.assembler.Emit(arm64.BCondLabel(arm64.OpBEQ, paths[idx]))
+	}
+	ctx.assembler.Emit(arm64.BLabel(paths[count]))
+	for idx, label := range paths {
+		ctx.assembler.Bind(label)
+		if !l.emitEdge(ctx, targets[idx], from, labels, index) {
+			return false
+		}
+	}
+	return true
+}
+
+func (l arm64Lowerer) emitEdge(ctx *lowering, target, from int, labels []asm.Label, index map[int]int) bool {
+	if target == len(ctx.frame().code) {
+		if ctx.addr == 0 {
+			return l.complete(ctx)
+		}
+		return l.ret(ctx)
+	}
+	idx, ok := index[target]
+	if !ok {
+		return false
+	}
+	if target > from {
+		ctx.assembler.Emit(arm64.BLabel(labels[idx]))
+		return true
+	}
+
+	a := ctx.assembler
+	vCtrl := ctx.pin(scratchCtrl)
+	budget := a.Reg(asm.RegTypeInt, asm.Width64)
+	a.Emit(arm64.LDR(budget, vCtrl, int16(journalBudget*8)))
+	a.Emit(arm64.SUBI(budget, budget, 1))
+	a.Emit(arm64.STR(budget, vCtrl, int16(journalBudget*8)))
+	a.Emit(arm64.CBNZLabel(budget, labels[idx]))
+	l.trapFlushed(ctx, trapYield, target)
+	return true
 }
