--- interp/jit_arm64_cfg_test.go
@@ -1,185 +1,0 @@
-package interp
-
-import (
-	"context"
-	"runtime"
-	"testing"
-
-	"github.com/siyul-park/minivm/instr"
-	"github.com/siyul-park/minivm/program"
-	"github.com/siyul-park/minivm/types"
-	"github.com/stretchr/testify/require"
-)
-
-// TestCompiler_CompileCFG covers Phase 2 of the whole-CFG JIT: compiling a
-// straight-line, single-basic-block, RETURN-terminated function directly
-// (without a recorded trace) into one native callable, and rejecting every
-// function shape Phase 2 does not yet support.
-func TestCompiler_CompileCFG(t *testing.T) {
-	if runtime.GOARCH != "arm64" {
-		t.Skip("native JIT is only available on arm64")
-	}
-
-	t.Run("straight-line arithmetic function compiles and matches threaded execution", func(t *testing.T) {
-		// (a + b) * 2, exercising I32_ADD, I32_CONST, and I32_MUL — all within
-		// lowerCFG's static coverage — inside a single RETURN-terminated block.
-		callee := types.NewFunctionBuilder(&types.FunctionType{
-			Params:  []types.Type{types.TypeI32, types.TypeI32},
-			Returns: []types.Type{types.TypeI32},
-		}).Emit(
-			instr.New(instr.LOCAL_GET, 0),
-			instr.New(instr.LOCAL_GET, 1),
-			instr.New(instr.I32_ADD),
-			instr.New(instr.I32_CONST, 2),
-			instr.New(instr.I32_MUL),
-			instr.New(instr.RETURN),
-		).MustBuild()
-
-		b := program.NewBuilder()
-		b.Globals(types.TypeI32)
-		idx := b.Const(callee)
-		// CALL pops the callee off the top of the stack, so the ref goes last:
-		// args first (in declared param order), then CONST_GET of the function.
-		b.Emit(instr.I32_CONST, 3).
-			Emit(instr.I32_CONST, 4).
-			Emit(instr.CONST_GET, uint64(idx)).
-			Emit(instr.CALL).
-			Emit(instr.GLOBAL_SET, 0)
-		prog, err := b.Build()
-		require.NoError(t, err)
-
-		i := New(prog, WithThreshold(-1))
-		defer i.Close()
-
-		c, err := newCompiler()
-		require.NoError(t, err)
-		defer c.Close()
-
-		addr := int(i.constants[idx].Ref())
-		fn, ok := i.function(addr)
-		require.True(t, ok)
-
-		mod, ok, err := c.compileCFG(i, addr, fn)
-		require.NoError(t, err)
-		require.True(t, ok)
-		i.install(mod, false)
-
-		require.NoError(t, i.Run(context.Background()))
-		got, err := i.Global(0)
-		require.NoError(t, err)
-		require.Equal(t, int32(14), got.I32())
-	})
-
-	t.Run("multi-block function compiles", func(t *testing.T) {
-		b := types.NewFunctionBuilder(&types.FunctionType{
-			Params:  []types.Type{types.TypeI32},
-			Returns: []types.Type{types.TypeI32},
-		})
-		alt := b.Label()
-		b.Emit(instr.New(instr.LOCAL_GET, 0)).
-			BrIf(alt).
-			Emit(instr.New(instr.I32_CONST, 1)).
-			Emit(instr.New(instr.RETURN)).
-			Bind(alt).
-			Emit(instr.New(instr.I32_CONST, 2)).
-			Emit(instr.New(instr.RETURN))
-		fn := b.MustBuild()
-
-		i := New(program.New(nil))
-		defer i.Close()
-
-		c, err := newCompiler()
-		require.NoError(t, err)
-		defer c.Close()
-
-		_, ok, err := c.compileCFG(i, 1, fn)
-		require.NoError(t, err)
-		require.True(t, ok)
-	})
-
-	t.Run("branches and loops match threaded execution", func(t *testing.T) {
-		calleeBuilder := types.NewFunctionBuilder(&types.FunctionType{
-			Params:  []types.Type{types.TypeI32},
-			Returns: []types.Type{types.TypeI32},
-		}).WithLocals(types.TypeI32)
-		loop := calleeBuilder.Label()
-		done := calleeBuilder.Label()
-		calleeBuilder.Emit(instr.New(instr.I32_CONST, 0)).
-			Emit(instr.New(instr.LOCAL_SET, 1)).
-			Bind(loop).
-			Emit(instr.New(instr.LOCAL_GET, 0)).
-			Emit(instr.New(instr.I32_EQZ)).
-			BrIf(done).
-			Emit(instr.New(instr.LOCAL_GET, 1)).
-			Emit(instr.New(instr.LOCAL_GET, 0)).
-			Emit(instr.New(instr.I32_ADD)).
-			Emit(instr.New(instr.LOCAL_SET, 1)).
-			Emit(instr.New(instr.LOCAL_GET, 0)).
-			Emit(instr.New(instr.I32_CONST, 1)).
-			Emit(instr.New(instr.I32_SUB)).
-			Emit(instr.New(instr.LOCAL_SET, 0)).
-			Br(loop).
-			Bind(done).
-			Emit(instr.New(instr.LOCAL_GET, 1)).
-			Emit(instr.New(instr.RETURN))
-		callee := calleeBuilder.MustBuild()
-
-		b := program.NewBuilder()
-		b.Globals(types.TypeI32)
-		idx := b.Const(callee)
-		b.Emit(instr.I32_CONST, 5).
-			Emit(instr.CONST_GET, uint64(idx)).
-			Emit(instr.CALL).
-			Emit(instr.GLOBAL_SET, 0)
-		prog, err := b.Build()
-		require.NoError(t, err)
-
-		threaded := New(prog, WithThreshold(-1))
-		defer threaded.Close()
-		require.NoError(t, threaded.Run(context.Background()))
-		want, err := threaded.Global(0)
-		require.NoError(t, err)
-
-		jit := New(prog, WithThreshold(-1))
-		defer jit.Close()
-		c, err := newCompiler()
-		require.NoError(t, err)
-		defer c.Close()
-		addr := int(jit.constants[idx].Ref())
-		fn, ok := jit.function(addr)
-		require.True(t, ok)
-		mod, ok, err := c.compileCFG(jit, addr, fn)
-		require.NoError(t, err)
-		require.True(t, ok)
-		jit.install(mod, false)
-		require.NoError(t, jit.Run(context.Background()))
-		got, err := jit.Global(0)
-		require.NoError(t, err)
-		require.Equal(t, want, got)
-	})
-
-	t.Run("unsupported opcode compiles an exact fallback", func(t *testing.T) {
-		// I32_DIV_S needs runtime trap semantics the baseline lowerer does not
-		// duplicate, so the CFG exits at that opcode and threaded dispatch owns it.
-		fn := types.NewFunctionBuilder(&types.FunctionType{
-			Params:  []types.Type{types.TypeI32, types.TypeI32},
-			Returns: []types.Type{types.TypeI32},
-		}).Emit(
-			instr.New(instr.LOCAL_GET, 0),
-			instr.New(instr.LOCAL_GET, 1),
-			instr.New(instr.I32_DIV_S),
-			instr.New(instr.RETURN),
-		).MustBuild()
-
-		i := New(program.New(nil))
-		defer i.Close()
-
-		c, err := newCompiler()
-		require.NoError(t, err)
-		defer c.Close()
-
-		_, ok, err := c.compileCFG(i, 1, fn)
-		require.NoError(t, err)
-		require.True(t, ok)
-	})
-}
