+++ interp/jit_plan.go
@@ -1,6 +1,13 @@
 package interp
 
-import "github.com/siyul-park/minivm/types"
+import (
+	"sort"
+
+	"github.com/siyul-park/minivm/analysis"
+	"github.com/siyul-park/minivm/instr"
+	"github.com/siyul-park/minivm/pass"
+	"github.com/siyul-park/minivm/types"
+)
 
 type planner interface {
 	plan(*compileInput) ([]plan, error)
@@ -35,10 +42,13 @@ const (
 )
 
 type block struct {
-	offset int
-	state  state
-	steps  []step
-	term   terminator
+	address   int
+	offset    int
+	hits      int64
+	canonical bool
+	state     state
+	steps     []step
+	term      terminator
 }
 
 type state struct {
@@ -107,6 +117,12 @@ func (p plan) valid() bool {
 	if len(p.blocks) == 0 {
 		return false
 	}
+	canonical := p.blocks[0].canonical
+	for _, block := range p.blocks[1:] {
+		if block.canonical != canonical {
+			return false
+		}
+	}
 	switch p.entry.kind {
 	case entryFunction:
 		if p.entry.anchor.addr <= 0 || p.entry.anchor.ip != 0 {
@@ -123,22 +139,445 @@ func (p plan) valid() bool {
 	default:
 		return false
 	}
-	offsets := make(map[int]struct{}, len(p.blocks))
+	offsets := make(map[anchor]struct{}, len(p.blocks))
 	for _, block := range p.blocks {
-		if _, ok := offsets[block.offset]; ok {
+		addr := block.address
+		if addr == 0 && p.entry.anchor.addr != 0 {
+			addr = p.entry.anchor.addr
+		}
+		a := anchor{addr: addr, ip: block.offset}
+		if _, ok := offsets[a]; ok {
 			return false
 		}
-		offsets[block.offset] = struct{}{}
+		offsets[a] = struct{}{}
 	}
-	if _, ok := offsets[p.entry.anchor.ip]; !ok {
+	if _, ok := offsets[p.entry.anchor]; !ok {
 		return false
 	}
 	for _, block := range p.blocks {
+		addr := block.address
+		if addr == 0 && p.entry.anchor.addr != 0 {
+			addr = p.entry.anchor.addr
+		}
 		for _, target := range block.term.targets {
-			if _, ok := offsets[target]; !ok {
+			if _, ok := offsets[anchor{addr: addr, ip: target}]; !ok {
 				return false
 			}
 		}
 	}
 	return true
+}
+
+type staticPlanner struct{}
+
+func (staticPlanner) plan(input *compileInput) ([]plan, error) {
+	if input == nil || input.function == nil || len(input.function.Code) == 0 || input.installed {
+		return nil, nil
+	}
+	if input.address == 0 {
+		for ip := 0; ip < len(input.function.Code); {
+			inst := instr.Instruction(input.function.Code[ip:])
+			if inst.Opcode() == instr.CALL || inst.Opcode() == instr.RETURN_CALL {
+				return nil, nil
+			}
+			ip += inst.Width()
+		}
+	}
+
+	manager := pass.NewManager()
+	pass.Register[*types.Function, []*analysis.BasicBlock](manager, analysis.NewBasicBlocksAnalysis())
+	blocks, err := pass.GetResult[[]*analysis.BasicBlock](manager, input.function)
+	if err != nil {
+		return nil, err
+	}
+	constants, heap := []types.Boxed(nil), []types.Value(nil)
+	if input.interpreter != nil {
+		constants, heap = input.interpreter.constants, input.interpreter.heap
+	}
+	facts, ok := planStates(input.function, blocks, constants, input.globals, heap)
+	if !ok {
+		return nil, nil
+	}
+
+	entryType := entryFunction
+	if input.address == 0 {
+		entryType = entryModule
+	}
+	result := plan{entry: entry{anchor: anchor{addr: input.address}, kind: entryType}}
+	result.blocks = make([]block, len(blocks))
+	for idx, source := range blocks {
+		target := block{address: input.address, offset: source.Start, canonical: true}
+		target.state.slots = make([]slot, len(facts[idx]))
+		copy(target.state.slots, facts[idx])
+		for ip := source.Start; ip < source.End; {
+			inst := instr.Instruction(input.function.Code[ip:])
+			next := ip + inst.Width()
+			step := step{op: inst.Opcode(), fn: input.address, ip: ip}
+			if inst.Opcode() == instr.CONST_GET {
+				constant := int(inst.Operand(0))
+				if constant < len(constants) && constants[constant].Kind() == types.KindRef {
+					step.ref = constants[constant].Ref()
+					step.known = true
+				}
+			}
+			switch inst.Opcode() {
+			case instr.BR:
+				target.term = terminator{kind: terminateBranch, ip: ip, targets: instr.Targets(input.function.Code, ip)}
+			case instr.BR_IF:
+				target.term = terminator{kind: terminateBranchIf, ip: ip, targets: append(instr.Targets(input.function.Code, ip), next)}
+			case instr.BR_TABLE:
+				target.term = terminator{kind: terminateBranchTable, ip: ip, targets: instr.Targets(input.function.Code, ip)}
+			case instr.RETURN:
+				target.term = terminator{kind: terminateReturn, ip: ip}
+			default:
+				target.steps = append(target.steps, step)
+			}
+			ip = next
+		}
+		if target.term.kind == terminateFallthrough {
+			if source.End == len(input.function.Code) {
+				if input.address == 0 {
+					target.term = terminator{kind: terminateComplete, ip: source.End}
+				} else {
+					target.term = terminator{kind: terminateReturn, ip: source.End}
+				}
+			} else {
+				target.term = terminator{kind: terminateBranch, ip: source.End, targets: []int{source.End}}
+			}
+		}
+		result.blocks[idx] = target
+	}
+	if !result.valid() {
+		return nil, nil
+	}
+	return []plan{result}, nil
+}
+
+type tracePlanner struct{}
+
+func (tracePlanner) plan(input *compileInput) ([]plan, error) {
+	if input == nil || input.interpreter == nil || input.interpreter.tracer == nil || input.function == nil {
+		return nil, nil
+	}
+	var plans []plan
+	for _, ip := range input.interpreter.tracer.anchors(input.address) {
+		a := anchor{addr: input.address, ip: ip}
+		tree := input.interpreter.tracer.rootAt(a)
+		if tree == nil || tree.root == nil || tree.root.kind == aborted {
+			continue
+		}
+		if (ip != 0) != (tree.root.kind == loop) {
+			continue
+		}
+		kind := entryFunction
+		if input.address == 0 {
+			kind = entryModule
+		}
+		if ip != 0 {
+			kind = entryLoop
+		}
+		planned := plan{entry: entry{anchor: a, kind: kind}}
+		planned.blocks = append(planned.blocks, traceBlock(tree.root, 0))
+		var continuations []block
+		for id, tr := range tree.branches {
+			if tr == nil || tr.kind == loop || tr.kind == aborted {
+				continue
+			}
+			hits := int64(0)
+			if id >= 0 && id < len(tree.hits) {
+				hits = tree.hits[id]
+			}
+			continuations = append(continuations, traceBlock(tr, hits))
+		}
+		sort.SliceStable(continuations, func(i, j int) bool {
+			if continuations[i].hits != continuations[j].hits {
+				return continuations[i].hits > continuations[j].hits
+			}
+			if continuations[i].address != continuations[j].address {
+				return continuations[i].address < continuations[j].address
+			}
+			return continuations[i].offset < continuations[j].offset
+		})
+		planned.blocks = append(planned.blocks, continuations...)
+		planned.spill = planSpill(planned.blocks)
+		if planned.valid() {
+			plans = append(plans, planned)
+		}
+	}
+	return plans, nil
+}
+
+func traceBlock(tr *trace, hits int64) block {
+	result := block{
+		address: tr.anchor.addr,
+		offset:  tr.anchor.ip,
+		hits:    hits,
+		steps:   append([]step(nil), tr.ops...),
+	}
+	switch tr.kind {
+	case returned:
+		result.term = terminator{kind: terminateReturn}
+	case completed:
+		result.term = terminator{kind: terminateComplete}
+	case partial:
+		resume := tr.anchor.ip
+		if len(tr.ops) > 0 {
+			resume = tr.ops[len(tr.ops)-1].target
+		}
+		result.term = terminator{kind: terminateFallback, ip: resume}
+	case loop:
+		result.term = terminator{kind: terminateBranch, ip: tr.anchor.ip, targets: []int{tr.anchor.ip}}
+	default:
+		result.term = terminator{kind: terminateFallback, ip: tr.anchor.ip}
+	}
+	return result
+}
+
+func planSpill(blocks []block) spillPolicy {
+	for _, block := range blocks {
+		if len(block.steps) == 0 {
+			continue
+		}
+		switch block.steps[len(block.steps)-1].op {
+		case instr.ARRAY_SET, instr.STRUCT_SET:
+			return spillForbidden
+		}
+	}
+	return spillAllowed
+}
+
+func planStates(fn *types.Function, blocks []*analysis.BasicBlock, constants []types.Boxed, globals []types.Kind, heap []types.Value) ([][]slot, bool) {
+	if len(fn.Handlers) > 0 {
+		return nil, false
+	}
+	if len(blocks) == 0 {
+		return nil, true
+	}
+	locals := localTypes(fn)
+	states := make([][]slot, len(blocks))
+	seen := make([]bool, len(blocks))
+	seen[0] = true
+	work := []int{0}
+	for len(work) > 0 {
+		idx := work[len(work)-1]
+		work = work[:len(work)-1]
+		state := append([]slot(nil), states[idx]...)
+		if !applyPlanBlock(fn, locals, constants, globals, heap, blocks[idx], &state) {
+			return nil, false
+		}
+		for _, succ := range blocks[idx].Succs {
+			if !seen[succ] {
+				seen[succ] = true
+				states[succ] = append([]slot(nil), state...)
+				work = append(work, succ)
+				continue
+			}
+			if len(states[succ]) != len(state) {
+				return nil, false
+			}
+			changed := false
+			for pos := range state {
+				merged, ok := mergeSlot(&states[succ][pos], state[pos])
+				if !ok {
+					return nil, false
+				}
+				changed = changed || merged
+			}
+			if changed {
+				work = append(work, succ)
+			}
+		}
+	}
+	for idx := range states {
+		if !seen[idx] {
+			return nil, false
+		}
+	}
+	return states, true
+}
+
+func mergeSlot(dst *slot, src slot) (bool, bool) {
+	if dst.kind != src.kind {
+		return false, false
+	}
+	changed := false
+	if dst.refKnown && (!src.refKnown || dst.ref != src.ref) {
+		dst.ref, dst.refKnown = 0, false
+		changed = true
+	}
+	if dst.calleeKnown && (!src.calleeKnown || dst.callee != src.callee) {
+		dst.callee, dst.calleeKnown = 0, false
+		changed = true
+	}
+	return changed, true
+}
+
+func applyPlanBlock(fn *types.Function, locals []types.Type, constants []types.Boxed, globals []types.Kind, heap []types.Value, block *analysis.BasicBlock, state *[]slot) bool {
+	for ip := block.Start; ip < block.End; {
+		inst := instr.Instruction(fn.Code[ip:])
+		if !applyPlanStep(fn, locals, constants, globals, heap, state, inst) {
+			return false
+		}
+		ip += inst.Width()
+	}
+	return true
+}
+
+func applyPlanStep(fn *types.Function, locals []types.Type, constants []types.Boxed, globals []types.Kind, heap []types.Value, state *[]slot, inst instr.Instruction) bool {
+	push := func(value slot) { *state = append(*state, value) }
+	pop := func(count int) bool {
+		if len(*state) < count {
+			return false
+		}
+		*state = (*state)[:len(*state)-count]
+		return true
+	}
+	switch inst.Opcode() {
+	case instr.NOP, instr.UNREACHABLE, instr.BR:
+		return true
+	case instr.LOCAL_GET:
+		idx := int(inst.Operand(0))
+		if idx >= len(locals) {
+			return false
+		}
+		push(slot{kind: locals[idx].Kind()})
+		return true
+	case instr.LOCAL_TEE:
+		return len(*state) > 0
+	case instr.UPVAL_GET:
+		idx := int(inst.Operand(0))
+		if idx >= len(fn.Captures) {
+			return false
+		}
+		push(slot{kind: fn.Captures[idx].Kind()})
+		return true
+	case instr.GLOBAL_GET:
+		idx := int(inst.Operand(0))
+		if idx >= len(globals) {
+			return false
+		}
+		push(slot{kind: globals[idx]})
+		return true
+	case instr.GLOBAL_TEE:
+		return len(*state) > 0
+	case instr.CONST_GET:
+		idx := int(inst.Operand(0))
+		if idx >= len(constants) {
+			return false
+		}
+		value := slot{kind: constants[idx].Kind()}
+		if value.kind == types.KindRef {
+			value.ref, value.refKnown = constants[idx].Ref(), true
+			if value.ref > 0 && value.ref < len(heap) {
+				if _, ok := heap[value.ref].(*types.Function); ok {
+					value.callee, value.calleeKnown = value.ref, true
+				}
+			}
+		}
+		push(value)
+		return true
+	case instr.DUP:
+		if len(*state) == 0 {
+			return false
+		}
+		push((*state)[len(*state)-1])
+		return true
+	case instr.SWAP:
+		if len(*state) < 2 {
+			return false
+		}
+		n := len(*state)
+		(*state)[n-1], (*state)[n-2] = (*state)[n-2], (*state)[n-1]
+		return true
+	case instr.SELECT:
+		if len(*state) < 3 {
+			return false
+		}
+		n := len(*state)
+		a, b := (*state)[n-2], (*state)[n-3]
+		if a.kind != b.kind {
+			return false
+		}
+		*state = (*state)[:n-3]
+		push(slot{kind: a.kind})
+		return true
+	case instr.ARRAY_GET:
+		if len(*state) < 2 {
+			return false
+		}
+		array := (*state)[len(*state)-2]
+		kind, ok := arrayKind(heap, array)
+		if !ok || !pop(2) {
+			return false
+		}
+		push(slot{kind: kind})
+		return true
+	case instr.CALL, instr.RETURN_CALL:
+		if len(*state) == 0 {
+			return false
+		}
+		callee := (*state)[len(*state)-1]
+		if !callee.calleeKnown || callee.callee <= 0 || callee.callee >= len(heap) {
+			return false
+		}
+		target, ok := heap[callee.callee].(*types.Function)
+		if !ok || target.Typ == nil || !pop(1+len(target.Typ.Params)) {
+			return false
+		}
+		if inst.Opcode() == instr.CALL {
+			for _, typ := range target.Typ.Returns {
+				push(slot{kind: typ.Kind()})
+			}
+		}
+		return true
+	case instr.RETURN:
+		returns := 0
+		if fn.Typ != nil {
+			returns = len(fn.Typ.Returns)
+		}
+		return len(*state) >= returns
+	case instr.STRUCT_NEW, instr.MAP_NEW, instr.CLOSURE_NEW:
+		return false
+	}
+	typ := inst.Type()
+	if typ.Pop == nil && typ.Push == nil || !pop(len(typ.Pop)) {
+		return false
+	}
+	for _, kind := range typ.Push {
+		if kind == instr.KindAny {
+			return false
+		}
+		push(slot{kind: types.Kind(kind)})
+	}
+	return true
+}
+
+func arrayKind(heap []types.Value, array slot) (types.Kind, bool) {
+	if !array.refKnown || array.ref <= 0 || array.ref >= len(heap) {
+		return 0, false
+	}
+	switch heap[array.ref].(type) {
+	case types.TypedArray[bool]:
+		return types.KindI1, true
+	case types.TypedArray[int8]:
+		return types.KindI8, true
+	case types.TypedArray[int32]:
+		return types.KindI32, true
+	case types.TypedArray[int64]:
+		return types.KindI64, true
+	case types.TypedArray[float32]:
+		return types.KindF32, true
+	case types.TypedArray[float64]:
+		return types.KindF64, true
+	default:
+		return 0, false
+	}
+}
+
+func localTypes(fn *types.Function) []types.Type {
+	var result []types.Type
+	if fn.Typ != nil {
+		result = append(result, fn.Typ.Params...)
+	}
+	return append(result, fn.Locals...)
 }
